# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Bont Sloots FC API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Player {
  id: number;
  name: string;
  position?: string | null;
  createdAt: string;
}

export type AwardType = (typeof AwardType)[keyof typeof AwardType];

export const AwardType = {
  mom: "mom",
  motm: "motm",
} as const;

export interface Award {
  id: number;
  playerId: number;
  playerName: string;
  fixtureId: number;
  fixtureOpponent: string;
  type: AwardType;
  createdAt: string;
}

export interface PlayerProfile {
  id: number;
  name: string;
  position?: string | null;
  createdAt: string;
  totalGoals: number;
  totalAssists: number;
  momCount: number;
  motmCount: number;
  awardHistory: Award[];
}

export interface PlayerStat {
  playerId: number;
  playerName: string;
  totalGoals: number;
  totalAssists: number;
  momCount: number;
  motmCount: number;
}

export interface CreatePlayerBody {
  name: string;
  position?: string | null;
}

export interface Fixture {
  id: number;
  opponent: string;
  matchDate: string;
  kickoffTime?: string | null;
  kickoffTbc: boolean;
  homeScore?: number | null;
  awayScore?: number | null;
  played: boolean;
  isHome: boolean;
  venue?: string | null;
  notes?: string | null;
}

export interface CreateFixtureBody {
  opponent: string;
  matchDate: string;
  kickoffTime?: string | null;
  kickoffTbc: boolean;
  isHome: boolean;
  venue?: string | null;
  notes?: string | null;
}

export interface UpdateFixtureBody {
  opponent?: string;
  matchDate?: string;
  kickoffTime?: string | null;
  kickoffTbc?: boolean;
  homeScore?: number | null;
  awayScore?: number | null;
  played?: boolean;
  isHome?: boolean;
  venue?: string | null;
  notes?: string | null;
}

export type StatEntryType = (typeof StatEntryType)[keyof typeof StatEntryType];

export const StatEntryType = {
  goal: "goal",
  assist: "assist",
} as const;

export interface StatEntry {
  id: number;
  playerId: number;
  fixtureId: number;
  type: StatEntryType;
  createdAt: string;
}

export type CreateStatBodyType =
  (typeof CreateStatBodyType)[keyof typeof CreateStatBodyType];

export const CreateStatBodyType = {
  goal: "goal",
  assist: "assist",
} as const;

export interface CreateStatBody {
  playerId: number;
  fixtureId: number;
  type: CreateStatBodyType;
}

export type CreateAwardBodyType =
  (typeof CreateAwardBodyType)[keyof typeof CreateAwardBodyType];

export const CreateAwardBodyType = {
  mom: "mom",
  motm: "motm",
} as const;

export interface CreateAwardBody {
  playerId: number;
  fixtureId: number;
  type: CreateAwardBodyType;
}

export type DashboardSeasonRecord = {
  played: number;
  wins: number;
  draws: number;
  losses: number;
  goalsFor: number;
  goalsAgainst: number;
};

export interface Dashboard {
  nextFixture?: Fixture | null;
  seasonRecord: DashboardSeasonRecord;
  topScorer?: PlayerStat | null;
  recentResults: Fixture[];
}

export interface AdminLoginBody {
  password: string;
}

export interface AdminLoginResult {
  success: boolean;
  token?: string | null;
}