# lib/api-zod/src/generated/types/adminLoginResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Bont Sloots FC API
 * OpenAPI spec version: 0.1.0
 */

export interface AdminLoginResult {
  success: boolean;
  token?: string | null;
}