# lib/api-zod/src/generated/types/dashboard.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Bont Sloots FC API
 * OpenAPI spec version: 0.1.0
 */
import type { DashboardSeasonRecord } from "./dashboardSeasonRecord";
import type { Fixture } from "./fixture";
import type { PlayerStat } from "./playerStat";

export interface Dashboard {
  nextFixture?: Fixture | null;
  seasonRecord: DashboardSeasonRecord;
  topScorer?: PlayerStat | null;
  recentResults: Fixture[];
}