# lib/api-zod/src/generated/types/fixture.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Bont Sloots FC API
 * OpenAPI spec version: 0.1.0
 */

export interface Fixture {
  id: number;
  opponent: string;
  matchDate: Date;
  kickoffTime?: string | null;
  kickoffTbc: boolean;
  homeScore?: number | null;
  awayScore?: number | null;
  played: boolean;
  isHome: boolean;
  venue?: string | null;
  notes?: string | null;
}