# lib/api-zod/src/generated/types/playerProfile.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Bont Sloots FC API
 * OpenAPI spec version: 0.1.0
 */
import type { Award } from "./award";

export interface PlayerProfile {
  id: number;
  name: string;
  position?: string | null;
  createdAt: Date;
  totalGoals: number;
  totalAssists: number;
  momCount: number;
  motmCount: number;
  awardHistory: Award[];
}