# lib/api-zod/src/generated/types/playerStat.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Bont Sloots FC API
 * OpenAPI spec version: 0.1.0
 */

export interface PlayerStat {
  playerId: number;
  playerName: string;
  totalGoals: number;
  totalAssists: number;
  momCount: number;
  motmCount: number;
}