# lib/api-zod/src/generated/types/statEntry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Bont Sloots FC API
 * OpenAPI spec version: 0.1.0
 */
import type { StatEntryType } from "./statEntryType";

export interface StatEntry {
  id: number;
  playerId: number;
  fixtureId: number;
  type: StatEntryType;
  createdAt: Date;
}